app/requirements.txt
Flask
tensorflow
numpy
Pillow
gunicorn